Hide tiny selection on mouse up
When you have an active selection and then click on the canvas, the
mouse down handler will render the selection, which currently has no
size. If you then release the click on the same location, the selection
stays as a single pixel, even after the selection is cleared. We must
re render the top layer to clear that.